+++ src/app/core/tree.py
@@ -57,11 +57,11 @@ class Tree:
 
     @logger.catch
     def proofFor(self, checksum):
-        """this method will not add a new checksum, but will check it exists and return the proof.
+        """this method will not add a new checksum, but will check it exists and return the proof. Assumes that the checksum has already been looked up and isn't in the cache.
         Errors: 
         ChecksumFormatError, ChecksumNotFoundError"""
 
-        return self.merkle.merkleProof({"checksum": checksum}).serialize()
+        return self.merkle.auditProof(checksum, commit=True).serialize()
 
     def validate(self, proof):
         """this method will return whether or not the proof submitted is valid. Assumes proof was generated by this service, with pymerkle.toJSONString(). Note the proof could be considered valid even if it's checksum isn't in this service's merkle tree.
