+++ src/tests/README.md
@@ -7,4 +7,4 @@ There are two kinds of tests we concern ourselves with at present:
 
 The Unit tests are found in the typical pytest locations and can all be run collectively via src/test.sh.
 
-The Stability Tests are tests of the performance of the service at the limits of its implementation. In our case, this is mainly done by stress testing the server. See src/stability_test.sh for how this is done. Briefly, we attempt to spam the server and use the logger to catch errors.
+The Stability Tests are tests of the performance of the service at the limits of its implementation. In our case, this is mainly done by stress testing the server. See src/stability_test.sh for how this is done. Briefly, we attempt to spam the server with various requests and use the logger to catch errors.
